device/requirements.txt
supabase>=2.0
python-dotenv>=1.0
pyserial>=3.5
